pyproject.toml
[tool.poetry]
name = "chat-bot"
version = "0.1.0"
description = ""
authors = ["t-yng <t.yng.greenmonkey@gmail.com>"]
readme = "README.md"
packages = [{include = "chat_bot"}]

[tool.poetry.dependencies]
python = "^3.11"
llama-index = "^0.7.9"
python-dotenv = "^1.0.0"
sentence-transformers = "^2.2.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
